Move room-related context components and hooks into src/rooms.tsx
